Relocate not shared Options component
Why:

----
We have a component in the components folder that is exclusively
used by the SelfAssessment options component. We would like to have
related components be co-located.

This commit:
----
- Combine `OptionSelect` and `Options/index` file
- Convert the Options/index file to Typescript
- Move the components/Options folder into SelfAssessment
- Temporary rename of `SelfAssessment/Option` =>
  `SelfAssessment/AssessmentOption`

Reviewers:
----
- This is a continued effort to convert the entire ecosystem into
typescript. Is an intermediary step into the overhaul that is going to
be the `SelfAssessment` folder.
